refactor: update imports from IUser to Models for consistency across modules
