+++ backend/src/modules/auth/services/FirebaseAuthService.ts
@@ -15,7 +15,7 @@ import admin from 'firebase-admin';
 import {UserRecord} from 'firebase-admin/lib/auth/user-record';
 import {applicationDefault} from 'firebase-admin/app';
 import {Inject, Service} from 'typedi';
-import {IUser} from 'shared/interfaces/IUser';
+import {IUser} from 'shared/interfaces/Models';
 import {IUserRepository} from 'shared/database';
 import {IAuthService} from '../interfaces/IAuthService';
 import {ChangePasswordBody, SignUpBody} from '../classes/validators';
